fix: resolved codebase consistency concerns
The codebase had 1 moderate and 1 high concern with the packages
in package-lock.json, which were resolved through an npm audit fix.
There were also concerns with the files makeing it to the repository
which simply needed to be git ignored to ensure there was a strong
foundation made for all future work with the rocket project.

Related Tickets: GEO-6