refactor: changed axios for fetch
This was mainly done for request memoization, which is automatically
supported by fetch and not axios. While axios has many benefits like
legacy support and simplified structure, fetch offers speed and
caching, making the user experience better. The convenience of axios
does not surpass the importance of user experience in terms of
frontend development, and there is no need for legacy support as this
is a newer project with the newest features.

GEO-28