docs: Add information to readme about env variables
There was no documentation associated to how the environment
variables are to be setup for local development. Therefore, a
short section was added to the readme to clarify how it should be
setup to run the application locally. This was tested by viewing
the result in WebStorm from Jetbrains (IDE).

GEO-21