rollback: Make the codebase a non-static export
The codebase was previously a static export because Firebase was
troublesome, but it seems to be working now. However, the recent
features of util classes and server actions did not work as the
client components need to be migrated first before server actions
can be used. This was tested with "npm run dev", "npm run build"
and "npm run start" while also performing a test preview channel
through the firebase command "firebase hosting:channel:deploy
test_deploy"