fix: Change actions to handle staging
The previous actions generated by Firebase made every PR have
a temporary release, but this should only happen when the PR
or the merge is towards staging. Therefore, this is a replica
of the main action but changes the branch and channelId. This
is about to be tested now through a PR and merge to staging.