refactor: Move launchlist APIs into util functions
There needs to be a mvoe towards server components where possible
to make space for server actions and server optimizations. This
was tested with an npm run dev to ensure the same features
modified were not removed. This involved printing to logs and
getting the same output as before.

GEO-22