refactor: Remove window event listerners from launch list
useEffects are powerful but prevent a component from being
a server component. Wherever possible, refactors are
important to move functionality away from useEffects
unless there is a strict purpose for a stateful function
that cannot be connected to component events or redirects.
This was tested by running an npm run dev and seeing if
moving the cursor over a rocket responded with movement.