refactor: Make Hover Item from Lauch List a Separate Component
The move to server components requires the lack of useEffects.
However, the hover item requires an interval to run to refresh
the time till launch. Therefore, it was separated into another
file to reduce the amount of code that had to be client
computed. This was supposed to be tested until my IP got
banned from the endpoint when leaving a previous version
running.