fix: Make Open Layers Client Side with no SSR
In dev mode, mapContainer would run "new Map" twice, causing
the first map to be blank while the second one rendered
below all the other components. To resolve this, the map
component is in the body of the function, and the import
of that component is done through a dynamic import with SSR
turned off. This was tested by running an npm run dev, to
see if the map appeared.