feat: Add api mocking with fake json data
Previously, the apis would contants lead to ip bans with too
many requests during development work and experimentation.
This now allows an environmental variable to be set and to
use temporary/expected data saved locally. This was tested
with npm run dev and checking if the fetch functions were
called.