refactor: Remove useState for selectedRocket
Previously, selectedRocket was handled in a stateful manner,
which did not allow for the use of server components. Not
only that, it required the use of useState and useEffect.
This was changed by adding a query parameter and handling it
on page load, which will become more efficient once there
is a switch to server components.

GEO-27