refactor: Add address url query parameter
Previously, the address was handled in a stateful manner, which
it still is, but it is now only set once after going through a
server action, which needs an await but thats only possible in
server components. This was the final preparation to make a
leap towards server components with the removal of setting
values with useStates.

GEO-26