feat: Add url query parameter for Address with debounce
Previously, address was handled through a useState, but now
it is through a query paraemter, which leaves room for
server components. During development, this led to issues
with repeated requests, slowing the platform down. This
was resolved with debouncing the requests, which should have
been done from the beginning. This was tested with an npm
run dev and checking if the results still appeared as the
same.

GEO-22