refactor: Change page.tsx to server component
Many previous commits were building up to changing the main page
into a server component. After several refactors of query
parameters and removing useState / useEffect, the use of server
components can finally be in use to add optimizations to all
processing and renderings. This was tested with an npm run dev.

GEO-25