refactor: Add launch list server component
Previously, launchlist was only client side while still using
server actions. This was changed by added a parent component
that is a server component to utilize all possible optimizations
and the stateful requirements of the hovered components was
moved to the children. This allows for optimizations to be placed
in the parent and passed down to the children if necessary. This
was tested with an npm run dev and checking for expected
functionality.

GEO-25