refactor: Change MapContainer to server component
This previously needed to fetch coordinates through props, but
after several commits, it called a server action. This allowed
a parent server component retrieve the values and pass it
through props to the new child MapContainer client component.
This was tested with an npm run dev and checking that the map
still loads.

GEO-25