feat: Add home page at root path
Previously, there was no welcome page for Geoship. Now, there is
at least a starting point with fixed routes and proper navigation
to the map. This was tested with an npm run dev and checking if
the root route loads and can navigate to the map by clicking the
button at the bottom of the screen.

GEO-20