feat: Add standardized data model for addresses
Previously, addresses were just using the 'any' type in typescript
which caused a lot of doubt in operations. This now ensures that
there are expected fields whenever the object is used and all the
errors associated to missing fields will be centered around the
address type parsing. This was tested by running an npm run dev,
looking for address suggestions, selecting an address, and
checking if there were any runtime errors.

GEO-23