feat: Add standardized models for launches and coordinates
Previously, there was the general use of 'any' for launches,
rockets, coordinates, etc. Now, there are standardized models
with specific parsing, so typescript and eslint can expect
types for both development and build detection, which also
isolates any data bugs to the parsing module. This was tested
with an npm run dev and checking that submitting an address,
hovering over a rocket, and selecting a rocket all work as
expected.

GEO-23