fix: Add empty if check to addressSuggestions
Previously, if the address was entered but then left empty, there
would be an error that there is no address. Now, there is a check
to see if the address is empty, and if it is, it returns an empty
list. This was tested with an npm run dev, entering an address,
and then deleting the address from the search bar to see if the
error would pop up.