feat: Add loading symbols to address suggestions
Previously, the use had no feedback if the address suggestions
were loading or were even going to pop up. Now, there is an
intederminant linear progress bar that shows the user that there
is something loading based on their typing. This was thoroughly
tested with an npm run dev and seeing if different rates of typing
and addresses would promote refreshing or glitchy results.

GEO-24