feat: Add server side function for detecting mobile
Previously, there was no way to distinguish mobile devices, which
prevented the frontend from changing between website and mobile
experience. Now, it is possible to detect it with a simple server
action that returns a boolean value, leaving room for more
functionality in the future. This was tested with an npm run dev
and adding a console log to the home page, checking the console
for a true/false dependent on chrome's developer tools.

GEO-17