feat: Add URL param retention with utils
Previously, any previous URL query parameters passed through would
be overwritten by the next redirect. Meaning, if an address was
previously selected, then a selected rocket would overwrte the
address input on the next redirect. Now, the previous parameters
are used for context if provided during the routing. This was
tested through an npm run dev and seeing if selecting an address
and a rocket would succesfully retain both.