feat: Retain Address input when used in url.
Previously, the address inputted would cause a redirect and the
address field would be left blank after loading. This has been
changed to display the address given from the URL query parameter
in the text field. This was tested with an npm run dev and checking
if the address appeared when it was and wasn't provided in the url.