fix: Change value for address to defaultValue
Previously, the URL query parameter was set to the value of the
textfield for address, which did not allow for changing the
value of the address. Now, the value is set to the inital default
value so it can be changed without issue from the user. This was
tested with an npm run dev, entering an address, reloading, and
then seeing if the textfield could be changed successfully.