refactor: Streamline RocketInfo
The coding layout of RocketInfo had repetitive calls to indeces
and it had some inline ternary operators that were hard to read.
Therefore, these were refactored to make readability better and
less indeces handling. This was tested by checking if RocketInfo
loaded with the same values when selecting a rocket.

GEO-37