feat: Add navbar
The layout was pretty basic and didn't have a consistent location
for static navigation to stay. Therefore, a basic navbar was
added that offers a place for a profile, title, and logo to be
housed. This was tested by checking if it renders and if it stays
present in both the main page and map page.