feat: Add smaller banner image
Google Adsense needed a banner, but the existing asset was too
large for ad use. Therefore, a new image was created using a size
adjuster to lower the kilobytes and used that for ad purposes.
This new asset was added to the public folder for future use but
is not required for ads to work (since the uploaded image is stored
in the google account rather than the codebase)