fix: Set calc height to children box
The height was set to 1, which led to weird expansion issues with
scrollable sections. This was fixed by calculating the exact size
it should be to prevent multiple scroll bars. Now, the children
can continue to flex to the parent's size as it is now fixed in
width and height.