feat: Add footer
The home page is under development, and the footer is just a small
step for a complete home screen. For now, it just has a link to
the map and a copy right statement. Since there are no other links
besides the map, the footer is pretty empty. This has only been
tested by the visual on the home page.