feat: Add Rocket Paginator to home page
This adds more sustenance to the home page. However, the clicking
to view other rockets feels a bit sluggish. The next version of
this component will most likely become a horizontally scrollable
section. Either that or the pagination needs to be smoother. This
was tested with renders with npm run dev.