feat: Add continuous scrolling to upcoming launches
Previously, the upcoming launches was limited to just three launches,
and it had some weird loading issues. A new UI was created to just
display all the upcoming launches at once, allowing the user to scroll.
This was tested by checking if it would render with an 'npm run dev'